Add files via upload
In today's society, various healthcare systems for the elderly exist, however, due to them living alone in the urbanized areas, a plethora of issues are faced by them when it comes to accessing the provided facilities. Few of them are :
1. Difficulty in Communication.
2. Lack of awareness about their health.
3. Lack of personal support in urban areas.
Therefore project Aarogya aims to empower the elderly(living alone) to lead a healthier lifestyle by creating a health monitoring system.
Major pillars of the project were Yoha, a voice assistant designed using Python and creation of the pipeline to successfully link the Mi band and Yoha to Google's Firebase, in order to successfully monitor the health in real-time. An android application was further created for extended support, fetching data from the maintained real-time database, and equipped with emergency buttons in event of an unfortunate incident such as illness and so on. Most of the reports were rendered in a visual format, such as using graphs and charts so that the elderly is always aware of his health scenario. Moreover, the voice assistant Yoha was specifically designed to pick keywords depicting an emergency( deteriorating health scenario), and immediately alert the loved ones as well as the nearby hospital.